+++ src/router/index.ts
@@ -1,6 +1,8 @@
 import { createRouter, createWebHistory } from 'vue-router'
 import Home from '@/views/Home.vue'
 import DefenseCalculators from '@/views/calculator/DefenseCalculators.vue'
+import CommunityMapsOverview from '@/views/community-maps/CommunityMapsOverview.vue'
+import CommunityMapsDetail from '@/views/community-maps/CommunityMapsDetail.vue'
 
 const routes = [
     {
@@ -12,6 +14,16 @@ const routes = [
         path: '/calculator/defense',
         name: 'calculator.defense',
         component: DefenseCalculators
+    },
+    {
+        path: '/community-maps',
+        name: 'community-maps',
+        component: CommunityMapsOverview
+    },
+    {
+        path: '/community-maps/:id',
+        name: 'community-maps.detail',
+        component: CommunityMapsDetail
     },
 ]